Remove testing code in GPUAcceleratedHSVPipe
